try to fix problems
