+++ src/index.html
@@ -65,6 +65,9 @@
         </p>
       </div>
     </div>
-    <script type="module" src="scripts/main.js"></script>
+    <script
+      type="module"
+      src="scripts/main.js"
+    ></script>
   </body>
 </html>
