+++ src/scripts/main.js
@@ -5,172 +5,26 @@
 // const game = new Game();
 
 // Write your code here
-export class Game {
-  constructor(initialState = null) {
-    this.size = 4;
-    this.score = 0;
-    this.status = 'start';
-    this.board = this.createEmptyBoard();
+import { Game } from '../modules/Game.class';
 
-    if (initialState) {
-      this.board = initialState;
-    } else {
-      this.addRandomTitle();
-      this.addRandomTitle();
-    }
-  }
+const game = new Game();
+const startButton = document.querySelector('button.start');
+const messageLose = document.querySelector('message-lose');
 
-  createEmptyBoard() {
-    return Array.from({ length: this.size }, () => Array(this.size).fill(0));
-  }
+startButton.addEventListener('click', () => {
+  messageLose.classList.add('hidden');
+  game.restart();
+  startButton.textContent = 'Restart';
+  startButton.classList.remove('start');
+  startButton.classList.add('restart');
+  renderBoard(game.getState());
+});
 
-  getState() {
-    return this.board;
-  }
+function renderBoard(board) {
+  const cells = document.querySelectorAll('.cell');
 
-  getScore() {
-    return this.score;
-  }
-
-  getStatus() {
-    return this.status;
-  }
-
-  start() {
-    this.board = this.createEmptyBoard;
-    this.score = 0;
-    this.status = 'playng';
-    this.addRandomTitle();
-    this.addRandomTitle();
-  }
-
-  restart() {
-    this.start();
-  }
-
-  addRandomTitle() {
-    const emptyCells = [];
-
-    for (let row = 0; row < this.size; row++) {
-      for (let col = 0; col < this.size; col++) {
-        if (this.board[row][col] === 0) {
-          emptyCells.push({ row, col });
-        }
-      }
-    }
-
-    if (emptyCells.length === 0) {
-      return;
-    }
-
-    const randomIndex = Math.floor(Math.random() * emptyCells.length);
-    const { row: randRow, col: randCol } = emptyCells[randomIndex];
-
-    this.board[randRow][randCol] = Math.random() < 0.9 ? 2 : 4;
-  }
-
-  moveToLeft() {
-    let moved = false;
-
-    for (let row = 0; row < this.size; row++) {
-      const newRow = this.compress(this.board[row]);
-      const mergedRow = this.merge(newRow);
-      const finalRow = this.compress(mergedRow);
-
-      if (this.board[row].join(',') !== finalRow.join(',')) {
-        moved = true;
-        this.board[row] = finalRow;
-      }
-    }
-
-    if (moved) {
-      this.addRandomTitle();
-      this.updateStatus();
-    }
-  }
-
-  moveToRight() {
-    this.board = this.board.map((row) => row.reverse());
-    this.moveToLeft();
-    this.board = this.board.map((row) => row.reverse());
-  }
-
-  moveUp() {
-    this.transpose();
-    this.moveToLeft();
-    this.transpose();
-  }
-
-  moveDown() {
-    this.transpose();
-    this.moveToRight();
-    this.transpose();
-  }
-
-  compress(row) {
-    return row
-      .filter((val) => val !== 0)
-      .contact(
-        Array(this.size - row.filter((val) => val !== 0).length).fill(0),
-      );
-  }
-
-  merge(row) {
-    for (let i = 0; i < this.size - 1; i++) {
-      if (row[i] !== 0 && row[i] === row[i] + 1) {
-        row[i] *= 2;
-        this.score += row[i];
-        row[i + 1] = 0;
-      }
-    }
-
-    return row;
-  }
-
-  transpose() {
-    const newBoard = this.createEmptyBoard();
-
-    for (let i = 0; i < this.size; i++) {
-      for (let j = 0; j < this.size; j++) {
-        newBoard[i][j] = this.board[j][i];
-      }
-    }
-    this.board = newBoard;
-  }
-
-  updateStatus() {
-    for (const row of this.board) {
-      for (const cell of row) {
-        if (cell === 2048) {
-          this.status = 'win';
-
-          return;
-        }
-      }
-    }
-
-    if (!this.canMove()) {
-      this.status = 'lose';
-    }
-  }
-
-  canMove() {
-    for (let i = 0; i < this.size; i++) {
-      for (let j = 0; j < this.size; j++) {
-        if (this.board[i][j] === 0) {
-          return true;
-        }
-
-        if (j < this.size - 1 && this.board[i][j] === this.board[i][j + 1]) {
-          return true;
-        }
-
-        if (i < this.size - 1 && this.board[i][j] === this.board[i + 1][j]) {
-          return true;
-        }
-      }
-    }
-
-    return false;
-  }
+  board.flat().forEach((value, index) => {
+    cells[index].textContent = value !== 0 ? value : '';
+    cells[index].className = `cell value - ${value}`;
+  });
 }
