+++ src/styles/main.scss
@@ -1,25 +1,31 @@
 body {
-  margin: 0;
   display: flex;
   align-items: center;
   justify-content: center;
+
   min-height: 100vh;
-  background: #fbf8ef;
+  margin: 0;
+
   font-family: sans-serif;
   font-size: 24px;
   font-weight: 900;
+
+  background: #fbf8ef;
 }
 
 .field-cell {
-  background: #d6cdc4;
+  user-select: none;
+
+  box-sizing: border-box;
   width: 75px;
   height: 75px;
   border-radius: 5px;
+
   color: #776e65;
-  box-sizing: border-box;
   text-align: center;
   vertical-align: center;
-  user-select: none;
+
+  background: #d6cdc4;
 
   &--2 {
     background: #eee4da;
@@ -30,78 +36,82 @@ body {
   }
 
   &--8 {
-    background: #f2b179;
     color: #f9f6f2;
+    background: #f2b179;
   }
 
   &--16 {
-    background: #f59563;
     color: #f9f6f2;
+    background: #f59563;
   }
 
   &--32 {
-    background: #f67c5f;
     color: #f9f6f2;
+    background: #f67c5f;
   }
 
   &--64 {
-    background: #f65e3b;
     color: #f9f6f2;
+    background: #f65e3b;
   }
 
   &--128 {
-    background: #edcf72;
     color: #f9f6f2;
+    background: #edcf72;
   }
 
   &--256 {
-    background: #edcc61;
     color: #f9f6f2;
+    background: #edcc61;
   }
 
   &--512 {
-    background: #edc850;
     color: #f9f6f2;
+    background: #edc850;
   }
 
   &--1024 {
-    background: #edc53f;
     color: #f9f6f2;
+    background: #edc53f;
   }
 
   &--2048 {
-    background: #edc22e;
     color: #f9f6f2;
+    background: #edc22e;
   }
 }
 
 .game-field {
-  background: #bbada0;
   border-spacing: 10px;
   border-radius: 5px;
+  background: #bbada0;
 }
 
 .game-header {
   display: flex;
-  width: 100%;
   justify-content: space-between;
+
+  box-sizing: border-box;
+  width: 100%;
   margin-bottom: 24px;
   padding: 10px;
-  box-sizing: border-box;
 }
 
 h1 {
-  background: #edc22e;
-  color: #f9f6f2;
-  width: 75px;
-  height: 75px;
-  font-size: 24px;
-  border-radius: 5px;
   display: flex;
   align-items: center;
   justify-content: center;
+
   box-sizing: border-box;
+  width: 75px;
+  height: 75px;
   margin: 0;
+  border-radius: 5px;
+
+  font-size: 24px;
+  color: #f9f6f2;
+
+  background: #edc22e;
 }
 
 .info {
@@ -109,14 +119,17 @@ h1 {
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  background: #d6cdc4;
+
+  box-sizing: border-box;
   width: 75px;
   height: 75px;
-  border-radius: 5px;
-  color: #776e65;
-  box-sizing: border-box;
-  font-size: 16px;
   margin: 0 8px 0 0;
+  border-radius: 5px;
+
+  font-size: 16px;
+  color: #776e65;
+
+  background: #d6cdc4;
 }
 
 .controls {
@@ -124,22 +137,24 @@ h1 {
 }
 
 .button {
-  border: none;
-  border-radius: 5px;
   cursor: pointer;
-  color: #f9f6f2;
-  font-family: sans-serif;
-  font-weight: 700;
-  font-size: 16px;
+
   width: 75px;
   height: 75px;
+  border: none;
+  border-radius: 5px;
+
+  font-family: sans-serif;
+  font-size: 16px;
+  font-weight: 700;
+  color: #f9f6f2;
 
   transition: 0.25s ease background;
 }
 
 .start {
-  background: #1dae28;
   font-size: 20px;
+  background: #1dae28;
 
   &:hover {
     background: #179921;
@@ -157,12 +172,14 @@ h1 {
 .message {
   box-sizing: border-box;
   width: 100%;
-  background: #d6cdc4;
-  color: #776e65;
   padding: 10px;
-  text-align: center;
   border-radius: 5px;
+
   font-size: 20px;
+  color: #776e65;
+  text-align: center;
+
+  background: #d6cdc4;
 }
 
 .hidden {
@@ -177,8 +194,8 @@ h1 {
 }
 
 .message-win {
-  background: #edc22e;
   color: #f9f6f2;
+  background: #edc22e;
 }
 
 .message-container {
